test-sort.js works properly renamed the main webcommand.js to index.js
