+++ test/test-sort.js
@@ -1,31 +1,20 @@
 var test = require("tap").test;
-var commandServer= require('../webcommand');
+var commandServer= require('../index');
 var stream = require('event-stream');
 
 test("make sure sort works", function (t) {
-	/*t.ok(true, 'TODO streams got my windows confused');
-	t.end();
-	*/
 	var commandServer=createCommandServer();
 	var iStream= stream.through();
 	var oStream= stream.through();
 	var err=false;
 	oStream.on('error', function(error){
 		err=true;
-		//throws error;
-		//t.ok(true, error);
-		//t.log(error);
-		//t.end();
 	});
 	oStream.on('end', function(){
 	    t.equal(err,false, "no error was emited");
 	    t.end();
 	});
-	//iStream.end();
-	commandServer.webCommand('wc',['-t', ',', '-k', '3', '-n', '-r'  ],iStream, oStream);
-	iStream.push('boo,*,23\nfoo,*,32\npoo,*,3\ndoo,*,2');
-	iStream.push(null);
-	iStream.end();
-	iStream.emit('close');
-	
+	commandServer.webCommand('sort',['-t', ',', '-k', '3', '-n', '-r'  ],iStream, oStream);
+	iStream.write('boo,*,23\nfoo,*,32\npoo,*,3\ndoo,*,2');
+	iStream.write(null);
 });
